+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed broken links in documentation.
+
 ## 1.11.0 - (September 27, 2022)
 
 * Added